:marty

+++ lib/cpp-test/array/linear_system_gtest.cpp
@@ -1,5 +1,6 @@
 // License: BSD 3 clause
 
+
 #define TICK_TEST_ROW_SIZE (10)
 #define TICK_TEST_COLUMN_SIZE (8)
 #define TICK_TEST_DATA_SIZE (100)
@@ -116,7 +117,7 @@ TYPED_TEST(LinearSystemTest, SymmetricLinearSystem) {
   // make it positive symmetric by computing its square B = A A'
   for (ulong i = 0; i < size; ++i) {
     for (ulong j = 0; j < size; ++j) {
-      for (ulong k = 0; k < size; ++k) {
+      for (size_t k = 0; k < size; ++k) {
         arrA(i, j) += sqrt_arrA(i, k) * sqrt_arrA(j, k);
       }
     }
