+++ lib/cpp/solver/sgd.cpp
@@ -7,7 +7,7 @@
 #include "tick/solver/sgd.h"
 
 template <class T, class K>
-TSGD<T, K>::TSGD(ulong epoch_size, T tol, RandType rand_type, T step, int record_every, int seed)
+TSGD<T, K>::TSGD(ulong epoch_size, T tol, RandType rand_type, T step, size_t record_every, int seed)
     : TStoSolver<T, K>(epoch_size, tol, rand_type, record_every, seed), step(step) {}
 
 template <class T, class K>
