+++ lib/include/tick/solver/sgd.h
@@ -30,7 +30,7 @@ class DLL_PUBLIC TSGD : public TStoSolver<T, K> {
 
  public:
   TSGD(ulong epoch_size = 0, T tol = 0., RandType rand_type = RandType::unif,
-       T step = 0.,  int record_every = 1, int seed = -1);
+       T step = 0.,  size_t record_every = 1, int seed = -1);
 
   inline T get_step_t() const { return step_t; }
